Fix typo: change redirectUri to redirectUrl (#27)
Example:
https://accounts.salla.sa/oauth2/auth?scope=offline_access&state=9a0e6714a39a8d39597a1c65220614f2&response_type=code&approval_prompt=auto&redirect_url=***&client_id=**

Error:
The request is missing a required parameter, includes an invalid parameter value, includes a parameter more than once, or is otherwise malformed. The 'redirect_url' parameter does not match any of the OAuth 2.0 Client's pre-registered redirect urls.

redirect_url is not a part of Salla's oauth2 parameters. The correct parameter is "redirect_uri".

Other option:
If "redirect_url" is the official parameter, as stated in the docs and other auth tutorials, it could be a bug in the auth endpoint.
https://accounts.salla.sa/oauth2/auth